+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\index.js":"1","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\reportWebVitals.js":"2","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\App.js":"3","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\components\\Main.jsx":"4","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\Styled-components\\MainStyle.js":"5","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\Assets\\Device.jsx":"6"},{"size":478,"mtime":1608045398016,"results":"7","hashOfConfig":"8"},{"size":364,"mtime":1608045398024,"results":"9","hashOfConfig":"8"},{"size":133,"mtime":1608045397972,"results":"10","hashOfConfig":"8"},{"size":5508,"mtime":1608047232048,"results":"11","hashOfConfig":"8"},{"size":2556,"mtime":1608047210677,"results":"12","hashOfConfig":"8"},{"size":377,"mtime":1608045397981,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1dh8nbu",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\index.js",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\reportWebVitals.js",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\App.js",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\components\\Main.jsx",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\Styled-components\\MainStyle.js",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\Assets\\Device.jsx",[]]
+[{"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\index.js":"1","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\reportWebVitals.js":"2","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\App.js":"3","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\components\\Main.jsx":"4","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\Styled-components\\MainStyle.js":"5","C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\Assets\\Device\\Device.jsx":"6"},{"size":478,"mtime":1608045398016,"results":"7","hashOfConfig":"8"},{"size":364,"mtime":1608045398024,"results":"9","hashOfConfig":"8"},{"size":133,"mtime":1608045397972,"results":"10","hashOfConfig":"8"},{"size":5508,"mtime":1608047232048,"results":"11","hashOfConfig":"8"},{"size":2563,"mtime":1608106473815,"results":"12","hashOfConfig":"8"},{"size":377,"mtime":1608106470322,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1dh8nbu",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\index.js",[],["27","28"],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\reportWebVitals.js",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\App.js",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\components\\Main.jsx",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\Styled-components\\MainStyle.js",[],"C:\\Users\\thoma\\desktop\\projet-favart\\lille-0920-js-p3-favart\\src\\Assets\\Device\\Device.jsx",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
